backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
email-validator==2.1.0
werkzeug==2.3.7